requirements.txt
accelerate==1.6.0
datasets==3.5.0
hydra-core==1.3.2
jsonlines==4.0.0
math-verify==0.7.0
numpy==1.26.4
peft==0.15.1
torch==2.6.0
torchaudio==2.6.0
torchdata==0.11.0
torchvision==0.21.0
vllm==0.8.5.post1
transformers==4.51.3